requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
python-telegram-bot==20.6
twilio==8.10.0
pandas==2.1.3
numpy==1.26.2
aiohttp==3.9.1
redis==5.0.1
jinja2==3.1.2
fastapi-socketio==0.0.10
python-multipart==0.0.6
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
beautifulsoup4==4.12.2
plotly==5.18.0
dash==2.14.1
scikit-learn==1.3.2
loguru==0.7.2
